+++ src/index.ts
@@ -1,3 +1,5 @@
 export { resolveJSONSchema, generateJoiStatement, formatJoi, Options, SubSchemas } from './joi';
 import * as Joi from './extendedJoi';
 export { Joi };
+export { AllOfSchema } from './extendedJoi/allOf';
+export { OneOfSchema } from './extendedJoi/oneOf';