+++ src/joi/index.ts
@@ -1,4 +1,4 @@
 export { resolveJSONSchema } from './resolve';
-export { Options, SubSchemas } from './options';
+export { ResolveOptions as Options, SubSchemas } from './options';
 export { generateJoiStatement } from './generate';
 export { formatJoi } from './format';