+++ src/joi/options.ts
@@ -5,10 +5,15 @@ export interface SubSchemas {
   [k: string]: JSONSchema4;
 }
 
-export interface Options {
+export interface ResolveOptions {
   subSchemas?: SubSchemas;
   rootSchema?: JSONSchema4;
   useDeprecatedJoi?: boolean;   // true: former 'joi' lib, false: '@hapi/joi' lib
-  useExtendedJoi?: boolean;     // only valid when `useDeprecatedJoi` is true,
-  // use extended Joi to support allOf and oneOf
+  useExtendedJoi?: boolean;     // use extended Joi (only when it is required to
+  // support allOf and oneOf using deprecatedJoi)
+}
+
+export interface FormatOptions {
+  joiName?: string;
+  extendedJoiName?: string;
 }